requirements.txt
Flask==3.0.3
flask-cors==4.0.1
openai==1.30.0
sentence-transformers==2.6.0
redis==5.0.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
numpy==1.26.4